- move helper functions back to circlepack.js - use pandemonium for array shuffling - use for...in instead of keys.forEach
