+++ index.d.ts
@@ -1,2 +1,3 @@
+export {default as circlepack} from './circlepack';
 export {default as circular} from './circular';
 export {default as random} from './random';
