+++ index.js
@@ -4,10 +4,11 @@
  *
  * Library endpoint.
  */
-var circular = require('./circular.js'),
-    circlepack = require('./circlepack.js'),
+var circlepack = require('./circlepack.js'),
+    circular = require('./circular.js'),
     random = require('./random.js');
 
+exports.circlepack = circlepack;
 exports.circular = circular;
 exports.random = random;
-exports.circlepack = circlepack;
+
