+++ package.json
@@ -51,6 +51,7 @@
   "dependencies": {
     "graphology-types": "^0.15.3",
     "graphology-utils": "^1.6.0",
-    "lodash": "^4.17.15"
+    "lodash": "^4.17.15",
+    "pandemonium": "^1.5.0"
   }
 }
